hydra-github-bridge: fix race condition that reports status to wrong commit
The race condition that this commit fixes:

1. commit A is pushed
2. build A starts
3. commit B is force pushed, updating the jobset's flake URL
4. build A finishes
5. status of build A is reported to commit B, as that's the jobset's flake URL

To fix this, instead of getting the flake URL from the jobset,
we instead get it from the latest evaluation that the job is a part of.
This is close to what Hydra does in its UI on the /build page for "Part of:", too.